change the word vendor
Hi the word "vendor" has been used inappropriately. It confuses a reader. This text needs to be readable for everyone. Why not to use the main folder since the author simply describes folders. In English vocabulary vendor means seller (it is a word derived from Latin language). The instruction becomes more communicative when one avoids use of the passive voice.